sharpened repo, small changes and created helpers folder
