+++ Stefan/.gitignore
@@ -1,2 +1,9 @@
 __pycache__
 *.vscode
+CMakeCache.txt
+*.cmake
+CMakeFiles
+Makefile
+Results-*/
+*.o
+
