--- Stefan/atvd3/.gitignore
@@ -1,10 +1,0 @@
-CMakeFiles
-.vscode
-media
-CMakeCache.txt
-*.cmake
-Makefile
-numero
-atvd3
-*.jpg
-*.jpeg
